@@ -1,205 +1,418 @@
 package org.example;
 
-import org.junit.Before;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.*;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.time.LocalDate;
-
-import static org.junit.Assert.*;
+import java.util.List;
+import java.util.Scanner;
 
 public class AppTest {
+
     private AuthSystem authSystem;
     private CarService carService;
-    private CarRentalSystem carRentalSystem;
+    private NotificationManager notificationManager;
+    private ReceiptGenerator receiptGenerator;
 
-    @Before
-    public void setUp() throws IOException {
-        resetUserFile(); // إعادة تعيين ملف المستخدمين قبل كل اختبار
-        authSystem = new AuthSystem(); // تهيئة النظام بعد إعادة التعيين
+    @BeforeEach
+    void setup() throws Exception {
+        // إعادة ضبط عداد السيارات
+        Car.resetIdCounter();
+
+        // تحضير ملف المستخدمين
+        File loginFile = new File("login.txt");
+        if (loginFile.exists()) loginFile.delete();
+        loginFile.createNewFile();
+
+        try (FileWriter writer = new FileWriter(loginFile)) {
+            writer.write("testuser,123,0500000000,test@example.com,Address\n");
+        }
+
+        // تحضير ملف السيارات
+        File carsFile = new File("cars.txt");
+        if (carsFile.exists()) carsFile.delete();
+        carsFile.createNewFile();
+
+        try (FileWriter carsWriter = new FileWriter(carsFile)) {
+            carsWriter.write("Sedan,Toyota,Camry,200.0,testuser,2024-12-09,2024-12-31,Standard\n");
+            carsWriter.write("SUV,Ford,Explorer,300.0,Ahmed,2024-12-05,2024-12-20,7\n");
+        }
+
+        authSystem = new AuthSystem();
         carService = new CarService();
-        carRentalSystem = CarRentalSystem.getInstance();
-    }
+        notificationManager = new NotificationManager();
+        receiptGenerator = new ReceiptGenerator();
 
-    // طريقة لإعادة تعيين ملف `login.txt`
-    private void resetUserFile() throws IOException {
-        File userFile = new File("login.txt");
-        if (userFile.exists()) {
-            FileWriter writer = new FileWriter(userFile, false); // الكتابة فوق الملف بمحتوى فارغ
-            writer.write(""); // تفريغ محتويات الملف
-            writer.close();
+        carService.loadCarsFromFile();
+        if (carService.getCarById(1) == null || carService.getCarById(2) == null) {
+            fail("Cars not loaded from file as expected.");
         }
     }
 
     // اختبارات AuthSystem
     @Test
-    public void testRegisterAndLogin() throws Exception {
-        authSystem.register("testUser", "password123", "1234567890", "test@example.com", "Test Address");
-        User user = authSystem.login("testUser", "password123");
+    void testLoginSuccess() {
+        User user = authSystem.login("testuser", "123");
         assertNotNull(user);
-        assertEquals("testUser", user.getUsername());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testRegisterWithDuplicatePhoneNumber() throws Exception {
-        authSystem.register("user1", "password123", "1234567890", "user1@example.com", "Address1");
-        authSystem.register("user2", "password456", "1234567890", "user2@example.com", "Address2"); // نفس رقم الهاتف
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testRegisterWithDuplicateEmail() throws Exception {
-        authSystem.register("user1", "password123", "1234567890", "user1@example.com", "Address1");
-        authSystem.register("user2", "password456", "0987654321", "user1@example.com", "Address2"); // نفس البريد الإلكتروني
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testLoginWithNonexistentUser() {
-        authSystem.login("nonexistentUser", "password123"); // تسجيل الدخول بمستخدم غير موجود
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testLoginWithWrongPassword() throws Exception {
-        authSystem.register("user1", "password123", "1234567890", "user1@example.com", "Address1");
-        authSystem.login("user1", "wrongPassword");
-    }
-
-    // اختبارات CarService
-    @Test
-    public void testAddCarWithValidDates() {
-        Car car = new Sedan("Toyota", "Camry", 150.0, "Standard", "Owner1",
-                LocalDate.of(2024, 1, 1), LocalDate.of(2024, 1, 10));
-        carService.addCar(car);
-        assertNotNull(carService.getCarById(car.getId()));
+        assertEquals("testuser", user.getUsername());
     }
 
     @Test
-    public void testGenerateReceipt() {
-        ReceiptGenerator receiptGenerator = new ReceiptGenerator();
-        Car car = new Sedan("Toyota", "Camry", 150.0, "Standard", "Owner1",
-                LocalDate.of(2024, 1, 1), LocalDate.of(2024, 1, 10));
-        LocalDate startDate = LocalDate.of(2024, 1, 2);
-        LocalDate endDate = LocalDate.of(2024, 1, 5);
-
-        // No assertion needed, just ensure no exceptions occur and receipt displays
-        receiptGenerator.generateReceipt(car, startDate, endDate);
+    void testLoginInvalidPassword() {
+        assertThrows(IllegalArgumentException.class, () -> authSystem.login("testuser", "wrong"));
     }
 
     @Test
-    public void testNotificationManager() {
-        NotificationManager notificationManager = new NotificationManager();
-        notificationManager.addNotification("user1", "Your car has been rented.");
-        notificationManager.addNotification("user1", "Another notification.");
-        notificationManager.addNotification("user2", "A different user's notification.");
-
-        // Capture the output
-        notificationManager.displayNotifications("user1"); // Ensure messages for user1 are displayed and cleared
-        notificationManager.displayNotifications("user2"); // Ensure messages for user2 are displayed and cleared
-    }
-
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testAddCarWithInvalidDates() {
-        Car car = new Sedan("Toyota", "Camry", 150.0, "Standard", "Owner1",
-                LocalDate.of(2024, 1, 10), LocalDate.of(2024, 1, 1)); // تواريخ غير صالحة
-        carService.addCar(car);
+    void testLoginUserNotFound() {
+        assertThrows(IllegalArgumentException.class, () -> authSystem.login("notfound", "123"));
     }
 
     @Test
-    public void testDisplayAvailableCarsExcludingOwner() {
-        Car car1 = new Sedan("Toyota", "Camry", 200.0, "Luxury", "Ali",
-                LocalDate.of(2024, 1, 1), LocalDate.of(2024, 1, 15));
-        Car car2 = new SUV("Nissan", "Patrol", 400.0, 5, "Ahmed",
-                LocalDate.of(2024, 1, 5), LocalDate.of(2024, 1, 20));
-        carService.addCar(car1);
-        carService.addCar(car2);
-
-        // يجب أن يعرض فقط سيارات أحمد
-        carService.displayAvailableCarsExcludingOwner("Ali");
+    void testRegisterNewUser() throws IOException {
+        authSystem.register("newuser", "pass", "0599999999", "new@example.com", "SomeAddress");
+        User user = authSystem.login("newuser", "pass");
+        assertNotNull(user);
+        assertEquals("newuser", user.getUsername());
     }
 
     @Test
-    public void testGetCarByNonexistentId() {
-        Car car = carService.getCarById(9999); // معرف غير موجود
-        assertNull(car); // يجب أن يكون null
+    void testRegisterDuplicateUsername() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            authSystem.register("testuser", "456", "0511111111", "another@example.com", "Addr");
+        });
+    }
+
+    @Test
+    void testRegisterDuplicatePhone() throws IOException {
+        authSystem.register("user2", "pass2", "0512345678", "user2@example.com", "Add");
+        assertThrows(IllegalArgumentException.class, () -> {
+            authSystem.register("user3", "pass3", "0512345678", "user3@example.com", "Add3");
+        });
+    }
+
+    @Test
+    void testRegisterDuplicateEmail() throws IOException {
+        authSystem.register("user4", "pass4", "0512345679", "dup@example.com", "Add");
+        assertThrows(IllegalArgumentException.class, () -> {
+            authSystem.register("user5", "pass5", "0512345680", "dup@example.com", "Add5");
+        });
+    }
+
+    @Test
+    void testAuthSystemEmptyFile() throws Exception {
+        File loginFile = new File("login.txt");
+        loginFile.delete();
+        loginFile.createNewFile();
+        authSystem = new AuthSystem();
+        assertThrows(IllegalArgumentException.class, () -> authSystem.login("testuser", "123"));
+    }
+
+    @Test
+    void testAuthSystemInvalidRegister() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            authSystem.register("invalidUser", "pass", "abc", "invalidEmail", "");
+        });
+    }
+
+    // CarFactory Tests
+    @Test
+    void testCarFactorySedan() {
+        Car car = CarFactory.getCar("Sedan", "Honda", "Accord", 250.0, "owner1", "Luxury",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,15));
+        assertTrue(car instanceof Sedan);
+    }
+
+    @Test
+    void testCarFactorySuv() {
+        Car car = CarFactory.getCar("SUV", "Ford", "Edge", 300.0, "owner2", "5",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,15));
+        assertTrue(car instanceof SUV);
+    }
+
+    @Test
+    void testCarFactoryInvalidType() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            CarFactory.getCar("Truck", "Ford", "F150", 400.0, "owner", "",
+                    LocalDate.of(2024,12,10), LocalDate.of(2024,12,15));
+        });
+    }
+
+    @Test
+    void testCarFactoryInvalidSuvSeats() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            CarFactory.getCar("SUV", "Toyota", "Rav4", 200.0, "owner", "-1",
+                    LocalDate.of(2024,12,10), LocalDate.of(2024,12,15));
+        });
+    }
+
+    @Test
+    void testCarFactoryCaseInsensitive() {
+        Car car = CarFactory.getCar("SeDaN", "Hyundai", "Elantra", 150.0, "ownerX", "Standard",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,15));
+        assertTrue(car instanceof Sedan);
+    }
+
+
+
+
+    @Test
+    void testCarSetPriceDayNegative() {
+        Car car = CarFactory.getCar("Sedan", "Kia", "Rio", 100.0, "ownerX", "Standard",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,20));
+        assertThrows(IllegalArgumentException.class, () -> car.setPricePerDay(-50));
+    }
+
+    @Test
+    void testCarSetAvailabilityInvalid() {
+        Car car = CarFactory.getCar("SUV","Jeep","Wrangler",200.0,"ownerY","4",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,20));
+        assertThrows(IllegalArgumentException.class, () -> {
+            car.setAvailability(LocalDate.of(2024,12,21), LocalDate.of(2024,12,20));
+        });
+    }
+
+    @Test
+    void testCarSetAvailabilitySameDay() {
+        Car car = CarFactory.getCar("Sedan","BMW","3Series",200.0,"ownerZ","Standard",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,20));
+        car.setAvailability(LocalDate.of(2024,12,15), LocalDate.of(2024,12,15));
+        assertEquals(LocalDate.of(2024,12,15), car.getAvailableFrom());
+        assertEquals(LocalDate.of(2024,12,15), car.getAvailableTo());
+    }
+
+    // CarService Tests
+    @Test
+    void testCarServiceLoad() {
+        Car car = carService.getCarById(1);
+        assertNotNull(car, "Car should be loaded from file");
+        assertEquals("Toyota", car.getMake());
+    }
+
+    @Test
+    void testCarServiceAddCar() {
+        String input = "Toyota\nCamry 2020\n250\nSedan\n2024-12-10\n2024-12-20\nStandard\n";
+        Scanner s = new Scanner(input);
+        carService.addCarFromInput(s, "testuser2");
+
+        Car newCar = carService.getCarById(3);
+        assertNotNull(newCar, "New car should be added");
+        assertEquals("Toyota", newCar.getMake());
+        assertEquals("testuser2", newCar.getOwner());
+    }
+
+    @Test
+    void testCarServiceAddCarInvalidMake() {
+        String input = "T0y0ta\nCamry 2020\n250\nSedan\n2024-12-10\n2024-12-20\nStandard\n";
+        Scanner s = new Scanner(input);
+        carService.addCarFromInput(s, "testuser3");
+        assertNull(carService.getCarById(3), "No car should be added due to invalid make");
+    }
+
+    @Test
+    void testCarServiceAddCarNegativePrice() {
+        String input = "Nissan\nAltima 2020\n-100\nSedan\n2024-12-10\n2024-12-20\nStandard\n";
+        Scanner s = new Scanner(input);
+        carService.addCarFromInput(s, "testuser4");
+        assertNull(carService.getCarById(3), "No car should be added due to negative price");
+    }
+
+    @Test
+    void testCarServiceSplitFullPeriod() {
+        Car car = carService.getCarById(1);
+        assertNotNull(car, "Car 1 should exist before full period rent");
+        carService.splitCarAvailability(car, car.getAvailableFrom(), car.getAvailableTo());
+        assertNull(carService.getCarById(1), "Car 1 should be removed after full period rent");
+    }
+
+    @Test
+    void testCarServiceSplitPartialPeriod() {
+        Car car = carService.getCarById(2);
+        assertNotNull(car, "Car 2 should exist before partial period rent");
+
+        LocalDate start = LocalDate.of(2024, 12, 10);
+        LocalDate end = LocalDate.of(2024, 12, 15);
+        carService.splitCarAvailability(car, start, end);
+
+        Car updatedCar = carService.getCarById(2);
+        assertNotNull(updatedCar, "Car 2 should still exist after partial rent");
+    }
+
+    @Test
+    void testCarServiceGetCarsByOwner() {
+        List<Car> ownerCars = carService.getCarsByOwner("testuser");
+        assertEquals(1, ownerCars.size(), "testuser should own 1 car");
+
+        List<Car> noOwnerCars = carService.getCarsByOwner("noOne");
+        assertTrue(noOwnerCars.isEmpty(), "noOne should not own any car");
+    }
+
+    @Test
+    void testCarServiceRemoveCar() {
+        Car car = carService.getCarById(1);
+        assertNotNull(car);
+        carService.removeCar(car);
+        assertNull(carService.getCarById(1), "Car should be removed");
+    }
+
+    @Test
+    void testCarServiceDisplayAvailableCarsExcludingOwner() {
+        carService.displayAvailableCarsExcludingOwner("testuser");
+        carService.displayAvailableCarsExcludingOwner("Ahmed");
+        carService.displayAvailableCarsExcludingOwner("noOne");
+    }
+
+    // إضافة اختبار splitCarAvailability مع تواريخ غير صالحة تماماً
+    @Test
+    void testCarServiceSplitInvalidPeriod() {
+        Car car = carService.getCarById(2);
+        assertNotNull(car);
+        // start بعد availableTo
+        LocalDate start = LocalDate.of(2025, 1, 1);
+        LocalDate end = LocalDate.of(2025, 1, 5);
+        assertThrows(IllegalArgumentException.class, () -> {
+            carService.splitCarAvailability(car, start, end);
+        });
+    }
+
+    @Test
+    void testCarServiceNoCarsDisplay() {
+        // إزالة كل السيارات
+        Car c1 = carService.getCarById(1);
+        Car c2 = carService.getCarById(2);
+        if (c1 != null) carService.removeCar(c1);
+        if (c2 != null) carService.removeCar(c2);
+        // الآن لا توجد سيارات
+        carService.displayAvailableCarsExcludingOwner("anyUser"); // لا استثناء
+    }
+
+    // NotificationManager Tests
+    @Test
+    void testNotificationManagerSingleNotification() {
+        notificationManager.addNotification("testuser", "Your car has been rented.");
+        notificationManager.displayNotifications("testuser");
+    }
+
+    @Test
+    void testNotificationManagerMultipleUsers() {
+        notificationManager.addNotification("userA", "Message A");
+        notificationManager.addNotification("userB", "Message B1");
+        notificationManager.addNotification("userB", "Message B2");
+
+        notificationManager.displayNotifications("userA");
+        notificationManager.displayNotifications("userB");
+        notificationManager.displayNotifications("userC");
+    }
+
+    @Test
+    void testNotificationManagerNoNotification() {
+        notificationManager.displayNotifications("unknownUser");
+    }
+
+    @Test
+    void testNotificationManagerClearAfterDisplay() {
+        notificationManager.addNotification("testuser", "Note 1");
+        notificationManager.addNotification("testuser", "Note 2");
+        notificationManager.displayNotifications("testuser");
+        // الآن لا إشعارات
+        notificationManager.displayNotifications("testuser");
+    }
+
+    // ReceiptGenerator Tests
+    @Test
+    void testReceiptGenerator() {
+        Car car = carService.getCarById(2);
+        assertNotNull(car);
+        LocalDate start = LocalDate.of(2024, 12, 10);
+        LocalDate end = LocalDate.of(2024, 12, 15);
+        receiptGenerator.generateReceipt(car, start, end);
+    }
+
+    @Test
+    void testReceiptGeneratorOneDayRental() {
+        Car car = carService.getCarById(2);
+        assertNotNull(car);
+        LocalDate sameDay = LocalDate.of(2024, 12, 10);
+        receiptGenerator.generateReceipt(car, sameDay, sameDay);
+    }
+
+    @Test
+    void testReceiptGeneratorZeroPrice() {
+        Car zeroPriceCar = CarFactory.getCar("Sedan", "Zero", "Price", 0.0, "ownerZ", "Standard",
+                LocalDate.of(2024,12,10), LocalDate.of(2024,12,11));
+        receiptGenerator.generateReceipt(zeroPriceCar, LocalDate.of(2024,12,10), LocalDate.of(2024,12,10));
+    }
+
+    // إضافة اختبار start > end في ReceiptGenerator (قد لا يكون منطقي لكن لزيادة التغطية)
+    @Test
+    void testReceiptGeneratorInvalidDates() {
+        Car car = carService.getCarById(2);
+        assertNotNull(car);
+        // إن نفذت receiptGenerator بدون فحص قد نغطي سطر زيادة
+        LocalDate start = LocalDate.of(2024,12,15);
+        LocalDate end = LocalDate.of(2024,12,10);
+        // لا يوجد استثناء معرف، لكن نحاول فقط تغطية حالة ما.
+        receiptGenerator.generateReceipt(car, start, end);
+    }
+
+    // User Tests
+    @Test
+    void testUserValid() {
+        User user = new User("u", "p", "0500000000", "email@test.com", "Address");
+        assertEquals("u", user.getUsername());
+        assertTrue(user.validatePassword("p"));
+        assertFalse(user.validatePassword("wrongPass")); // تغطية مسار كلمة مرور خاطئة
+        assertEquals("0500000000", user.getPhoneNumber());
+        assertEquals("email@test.com", user.getEmail());
+        assertEquals("Address", user.getAddress());
+        assertTrue(user.toFile().contains("u,p,0500000000,email@test.com,Address"));
+    }
+
+    @Test
+    void testUserInvalidPhone() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            new User("x", "p", "abc", "valid@test.com", "Addr");
+        });
+    }
+
+    @Test
+    void testUserInvalidEmail() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            new User("y", "p", "0501234567", "invalidEmail", "Addr");
+        });
+    }
+
+    @Test
+    void testUserEmptyUsername() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            new User("", "p", "0501234567", "email@t.com", "Addr");
+        });
+    }
+
+    @Test
+    void testUserEmptyAddress() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            new User("userX", "p", "0501234567", "mail@test.com", "");
+        });
     }
 
     // اختبارات CarRentalSystem
     @Test
-    public void testInitialize() {
-        carRentalSystem.initialize();
-        assertEquals(2, carRentalSystem.getCarService().getCarsByOwner("Ali").size()
-                + carRentalSystem.getCarService().getCarsByOwner("Ahmed").size());
-    }
-
-    @Test
-    public void testRentCarWithValidDates() {
-        Car car = new Sedan("Toyota", "Yaris", 100.0, "Standard", "Owner1",
-                LocalDate.of(2024, 1, 1), LocalDate.of(2024, 1, 15));
-        carService.addCar(car);
-
-        LocalDate startDate = LocalDate.of(2024, 1, 2);
-        LocalDate endDate = LocalDate.of(2024, 1, 5);
-
-        carService.splitCarAvailability(car, startDate, endDate);
-
-        // التحقق من توفر السيارة بعد الاستئجار
-        assertEquals(LocalDate.of(2024, 1, 6), car.getAvailableFrom());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testRentCarOutsideAvailability() {
-        Car car = new SUV("Ford", "Escape", 300.0, 7, "Owner2",
-                LocalDate.of(2024, 3, 1), LocalDate.of(2024, 3, 15));
-        carService.addCar(car);
-
-        carService.splitCarAvailability(car, LocalDate.of(2024, 2, 20), LocalDate.of(2024, 2, 25)); // فترة خارج التوفر
-    }
-
-    @Test
-    public void testRegisterWithDuplicateUsernameAndEmail() throws Exception {
-        authSystem.register("testUser", "password123", "1234567890", "test@example.com", "Test Address");
-        try {
-            authSystem.register("testUser", "password1234", "0987654321", "test@example.com", "Another Address");
-            fail("Expected IllegalArgumentException for duplicate username and email.");
-        } catch (IllegalArgumentException e) {
-            assertEquals("Username is already taken.", e.getMessage());
-        }
+    void testCarRentalSystemGetInstance() {
+        CarRentalSystem sys1 = CarRentalSystem.getInstance();
+        CarRentalSystem sys2 = CarRentalSystem.getInstance();
+        assertSame(sys1, sys2, "Should return the same instance");
+        assertNotNull(sys1.getCarService(), "getCarService() should not return null");
     }
 
 
     @Test
-    public void testGetCarByIdWithInvalidId() {
-        assertNull(carService.getCarById(9999)); // اختبار معرف غير موجود
+    void testCarRentalSystemInvalidMainChoice() {
+        // إدخال خيار غير صحيح مرتين ثم 3 للخروج
+        String input = "9\n9\n3\n";
+        System.setIn(new ByteArrayInputStream(input.getBytes()));
+        CarRentalSystem system = CarRentalSystem.getInstance();
+        system.start();
+        system.processUserInput(); // سيطبع "Invalid choice." حتى نختار 3 للخروج
     }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testSplitCarAvailabilityWithInvalidDates() {
-        Car car = new Sedan("Toyota", "Camry", 150.0, "Standard", "Owner1",
-                LocalDate.of(2024, 1, 1), LocalDate.of(2024, 1, 10));
-        carService.addCar(car);
-
-        // هذا يجب أن يرمي استثناء
-        carService.splitCarAvailability(car, LocalDate.of(2024, 1, 11), LocalDate.of(2024, 1, 5));
-    }
-
-
-    @Test
-    public void testDisplayAvailableCarsExcludingOwnerWithNoCars() {
-        // عندما لا توجد سيارات مملوكة للمالك الحالي
-        carService.displayAvailableCarsExcludingOwner("NonexistentOwner");
-    }
-
-    @Test
-    public void testInitializeWithNoCars() {
-        carRentalSystem.initialize();
-        assertTrue(carRentalSystem.getCarService().getCarsByOwner("NonexistentOwner").isEmpty());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testRegisterWithInvalidPhoneNumber() throws Exception {
-        authSystem.register("invalidPhoneUser", "password123", "123", "user@example.com", "Address"); // رقم غير صالح
-    }
-
 }
