@@ -1,4 +1,3 @@
-
 package org.example;
 
 import java.time.LocalDate;
@@ -18,6 +17,7 @@ public class CarRentalSystem {
         this.notificationManager = new NotificationManager();
         this.receiptGenerator = new ReceiptGenerator();
         this.scanner = new Scanner(System.in);
+        carService.loadCarsFromFile();
     }
 
     public static CarRentalSystem getInstance() {
@@ -27,21 +27,14 @@ public class CarRentalSystem {
         return instance;
     }
 
-    public void initialize() {
-        carService.addCar(new Sedan("Toyota", "Camry", 200.0, "Standard", "Ali", LocalDate.of(2024, 12, 9), LocalDate.of(2024, 12, 31)));
-        carService.addCar(new SUV("Ford", "Explorer", 300.0, 7, "Ahmed", LocalDate.of(2024, 12, 5), LocalDate.of(2024, 12, 20)));
-    }
-
     public void start() {
         System.out.println("Welcome to the Wheelzy Rental System!");
     }
 
     public void processUserInput() {
         while (true) {
-            System.out.println("\n1. Login\n2. Register\n3. Exit");
-            System.out.print("Choose an option: ");
-            int choice = scanner.nextInt();
-            scanner.nextLine();
+            printMainMenu();
+            int choice = readIntFromScanner();
             switch (choice) {
                 case 1 -> handleLogin();
                 case 2 -> handleRegister();
@@ -54,14 +47,24 @@ public class CarRentalSystem {
         }
     }
 
-    public void handleLogin() {
-        System.out.print("Enter username: ");
-        String username = scanner.nextLine();
-        System.out.print("Enter password: ");
-        String password = scanner.nextLine();
+    private void printMainMenu() {
+        System.out.println("\n1. Login\n2. Register\n3. Exit");
+        System.out.print("Choose an option: ");
+    }
+
+    private int readIntFromScanner() {
+        int val = scanner.nextInt();
+        scanner.nextLine();
+        return val;
+    }
+
+    private void handleLogin() {
+        String username = promptInput("Enter username: ");
+        String password = promptInput("Enter password: ");
+
         try {
             User user = authSystem.login(username, password);
-            System.out.println("Welcome to Wheelzy , "+ username);
+            System.out.println("Welcome to Wheelzy , " + username);
             notifyOwnerOnLogin(user);
             userMenu(user);
         } catch (IllegalArgumentException e) {
@@ -69,21 +72,13 @@ public class CarRentalSystem {
         }
     }
 
-    private void notifyOwnerOnLogin(User user) {
-        notificationManager.displayNotifications(user.getUsername());
-    }
+    private void handleRegister() {
+        String username = promptInput("Enter username: ");
+        String password = promptInput("Enter password: ");
+        String phone = promptInput("Enter phone: ");
+        String email = promptInput("Enter email: ");
+        String address = promptInput("Enter address: ");
 
-    public void handleRegister() {
-        System.out.print("Enter username: ");
-        String username = scanner.nextLine();
-        System.out.print("Enter password: ");
-        String password = scanner.nextLine();
-        System.out.print("Enter phone: ");
-        String phone = scanner.nextLine();
-        System.out.print("Enter email: ");
-        String email = scanner.nextLine();
-        System.out.print("Enter address: ");
-        String address = scanner.nextLine();
         try {
             authSystem.register(username, password, phone, email, address);
         } catch (Exception e) {
@@ -91,12 +86,19 @@ public class CarRentalSystem {
         }
     }
 
+    private String promptInput(String message) {
+        System.out.print(message);
+        return scanner.nextLine();
+    }
+
+    private void notifyOwnerOnLogin(User user) {
+        notificationManager.displayNotifications(user.getUsername());
+    }
+
     public void userMenu(User user) {
         while (true) {
-            System.out.println("\n1. View and Rent Cars\n2. Add Your Car\n3. Logout\n0. Exit");
-            System.out.print("Choose an option: ");
-            int choice = scanner.nextInt();
-            scanner.nextLine();
+            printUserMenu();
+            int choice = readIntFromScanner();
             switch (choice) {
                 case 1 -> rentCar(user);
                 case 2 -> carService.addCarFromInput(scanner, user.getUsername());
@@ -104,93 +106,114 @@ public class CarRentalSystem {
                     System.out.println("Logging out...");
                     return;
                 }
-                case 0 -> {
-                    System.out.println("Exiting program...");
-                    System.exit(0);
-                }
+                case 0 -> exitProgram();
                 default -> System.out.println("Invalid choice.");
             }
         }
     }
 
+    private void printUserMenu() {
+        System.out.println("\n1. View and Rent Cars");
+        System.out.println("2. Add Your Car");
+        System.out.println("3. Logout");
+        System.out.println("0. Exit");
+        System.out.print("Choose an option: ");
+    }
+
+    private void exitProgram() {
+        System.out.println("Exiting program...");
+        System.exit(0);
+    }
+
     public void rentCar(User user) {
-        carService.displayAvailableCarsExcludingOwner(user.getUsername());
-        System.out.print("\nDo you want to rent a car? (yes/no): ");
-        String response = scanner.nextLine().trim().toLowerCase();
+        int availableCount = carService.displayAvailableCarsExcludingOwner(user.getUsername());
 
-        if (response.equals("yes")) {
-            System.out.print("Enter the ID of the car you want to rent: ");
-            int carId = scanner.nextInt();
-            scanner.nextLine(); // Consume newline
+        if (availableCount == 0) {
+            System.out.println("Returning to the menu...");
+            return;
+        }
 
-            Car car = carService.getCarById(carId);
-            if (car == null) {
-                System.out.println("Invalid car ID. Returning to the menu...");
-                return;
-            }
-
-            LocalDate startDate = null;
-            LocalDate endDate = null;
-
-            // Collect and validate rental start date
-            while (true) {
-                try {
-                    System.out.print("Enter rental start date (YYYY-MM-DD): ");
-                    startDate = LocalDate.parse(scanner.nextLine());
-                    if (startDate.isBefore(car.getAvailableFrom()) || startDate.isAfter(car.getAvailableTo())) {
-                        System.out.printf("Start date must be between %s and %s.%n",
-                                car.getAvailableFrom(), car.getAvailableTo());
-                        continue;
-                    }
-                    break;
-                } catch (Exception e) {
-                    System.out.println("Invalid date format. Please enter the date in YYYY-MM-DD format.");
-                }
-            }
-
-
-            while (true) {
-                try {
-                    System.out.print("Enter rental end date (YYYY-MM-DD): ");
-                    endDate = LocalDate.parse(scanner.nextLine());
-                    if (endDate.isBefore(startDate) || endDate.isAfter(car.getAvailableTo())) {
-                        System.out.printf("End date must be after %s and no later than %s.%n",
-                                startDate, car.getAvailableTo());
-                        continue;
-                    }
-                    break;
-                } catch (Exception e) {
-                    System.out.println("Invalid date format. Please enter the date in YYYY-MM-DD format.");
-                }
-            }
-
-
+        if (confirmAction("Do you want to rent a car? (yes/no): ")) {
+            int carId = getCarIdFromUser();
+            Car car = validateCarId(carId);
+            if (car == null) return;
+            LocalDate startDate = getRentalDate(car.getAvailableFrom(), car.getAvailableTo());
+            LocalDate endDate = getRentalEndDate(car, startDate);
             receiptGenerator.generateReceipt(car, startDate, endDate);
-
-
-            System.out.print("Do you want to confirm the rental? (yes/no): ");
-            String confirmation = scanner.nextLine().trim().toLowerCase();
-
-            if (confirmation.equals("yes")) {
-
-                carService.splitCarAvailability(car, startDate, endDate);
-
-
-                String owner = car.getOwner();
-                String message = String.format("Your car %s %s has been rented by %s from %s to %s.",
-                        car.getMake(), car.getModel(), user.getUsername(), startDate, endDate);
-                notificationManager.addNotification(owner, message);
-
-
-                System.out.printf("Car rented successfully from %s to %s.%n", startDate, endDate);
-                System.out.println("The courier will contact you via mobile phone number : "+user.getPhoneNumber()+" .");
-            } else {
-                System.out.println("Rental cancelled. Returning to the menu...");
-            }
+            confirmingRental(startDate, endDate, car, user);
         } else {
             System.out.println("Returning to the menu...");
         }
     }
+
+    private boolean confirmAction(String message) {
+        System.out.print(message);
+        String response = scanner.nextLine().trim().toLowerCase();
+        return response.equals("yes");
+    }
+
+    private int getCarIdFromUser() {
+        System.out.print("Enter the ID of the car you want to rent: ");
+        int val = scanner.nextInt();
+        scanner.nextLine();
+        return val;
+    }
+
+    private Car validateCarId(int carId) {
+        Car car = carService.getCarById(carId);
+        if (car == null) {
+            System.out.println("Invalid car ID. Returning to the menu...");
+        }
+        return car;
+    }
+
+    private LocalDate getRentalDate(LocalDate minDate, LocalDate maxDate) {
+        while (true) {
+            System.out.print("Enter rental start date (YYYY-MM-DD): ");
+            try {
+                LocalDate date = LocalDate.parse(scanner.nextLine());
+                if (date.isBefore(minDate) || date.isAfter(maxDate)) {
+                    System.out.printf("Date must be between %s and %s.%n", minDate, maxDate);
+                    continue;
+                }
+                return date;
+            } catch (Exception e) {
+                System.out.println("Invalid date format. Please enter the date in YYYY-MM-DD format.");
+            }
+        }
+    }
+
+    private LocalDate getRentalEndDate(Car car, LocalDate startDate) {
+        while (true) {
+            System.out.print("Enter rental end date (YYYY-MM-DD): ");
+            try {
+                LocalDate endDate = LocalDate.parse(scanner.nextLine());
+                if (endDate.isBefore(startDate) || endDate.isAfter(car.getAvailableTo())) {
+                    System.out.printf("End date must be after %s and no later than %s.%n", startDate, car.getAvailableTo());
+                    continue;
+                }
+                return endDate;
+            } catch (Exception e) {
+                System.out.println("Invalid date format. Please enter the date in YYYY-MM-DD format.");
+            }
+        }
+    }
+
+    private void confirmingRental(LocalDate startDate, LocalDate endDate, Car car, User user) {
+        if (confirmAction("Do you want to confirm the rental? (yes/no): ")) {
+            carService.splitCarAvailability(car, startDate, endDate);
+
+            String owner = car.getOwner();
+            String message = String.format("Your car %s %s has been rented by %s from %s to %s.",
+                    car.getMake(), car.getModel(), user.getUsername(), startDate, endDate);
+            notificationManager.addNotification(owner, message);
+            System.out.printf("Car rented successfully from %s to %s.%n", startDate, endDate);
+            System.out.println("The courier will contact you via mobile phone number : " + user.getPhoneNumber() + " .");
+        } else {
+            System.out.println("Rental cancelled. Returning to the menu...");
+        }
+    }
+
     public CarService getCarService() {
         return this.carService;
     }
