@@ -1,4 +1,3 @@
-
 package org.example;
 
 import java.time.LocalDate;
@@ -18,6 +17,11 @@ public class SUV extends Car {
     public void displayCarInfo() {
         System.out.printf("ID: %d | SUV: %s %s, Seats: %d, Price: %.2f SAR per Day%nAvailable From: %s To: %s%n",
                 id, make, model, numOfSeats, pricePerDay, availableFrom, availableTo);
+    }
+
+    @Override
+    protected String extraAttributes() {
+        return "," + numOfSeats;
     }
 
     public int getNumOfSeats() {
